bcr format, version check in samtools

+++ pipeline.sh
@@ -25,6 +25,11 @@
 # - bgzip
 # - vcftools
 
+#Byte size of EGSA fields
+GSAtext=4
+GSAsuff=1
+LCP=1
+
 WD=$(dirname $(readlink -f "$1"))
 
 #extract full path of working directory without trailing /
@@ -56,6 +61,8 @@ TIME_BWAIDX=${WD}/bwaindex.time
 TIME_BCFTOOLS=${WD}/bcftools.time
 REPORT_OUT=${WD}/report_${M}
 
+samtool_version=`samtools 2>&1 >/dev/null | grep Version | cut -d' ' -f 2 | cut -d. -f 1`
+
 echo "input 2: "${READS2}"."${EXT2}
 
 # 1.  Converts input reads to fasta, if not already in fasta -> reads1.fasta reads2.fasta 
@@ -84,11 +91,13 @@ if [ ! -f ${WD}/${READS1}.${READS2}.frc.fasta ]; then
 	rm ${WD}/${READS1}.rc.fasta ${WD}/${READS2}.rc.fasta
 fi
 
-# 3.  Builds EGSA -> reads1.reads2.frc.fasta.gesa
-
-if [ ! -f ${WD}/${READS1}.${READS2}.frc.fasta.gesa ]; then
+# 3.  If EGSA/BCR files do not exist (both), then builds EGSA -> reads1.reads2.frc.fasta.gesa
+if [[ ! -f ${WD}/${READS1}.${READS2}.frc.fasta.gesa && ! -f ${WD}/${READS1}.${READS2}.frc.fasta.out ]]; then
 	echo "building EGSA ..."
-	/usr/bin/time -v egsa -vvv -c ${WD}/${READS1}.${READS2}.frc.fasta 0 > ${TIME_EGSA} 2>&1
+	#/usr/bin/time -v egsa -vvv -c ${WD}/${READS1}.${READS2}.frc.fasta 0 > ${TIME_EGSA} 2>&1
+	/usr/bin/time -v gsufsort ${WD}/${READS1}.${READS2}.frc.fasta --gesa ${GSAtext} ${GSAsuff} ${LCP} > ${TIME_EGSA} 2>&1
+	mv ${WD}/${READS1}.${READS2}.frc.fasta.${GSAtext}.${GSAsuff}.${LCP}.1.gesa ${WD}/${READS1}.${READS2}.frc.fasta.gesa
+	
 	rm -rf ${WD}/tmp
 	rm -rf ${WD}/partition
 fi
@@ -97,14 +106,14 @@ fi
 
 if [ ! -f ${WD}/${READS1}.${READS2}.frc.fasta.clusters ]; then
 	echo "running eBWTclust ..."
-	/usr/bin/time -v eBWTclust -i ${WD}/${READS1}.${READS2}.frc.fasta > ${TIME_EBWTCLUST} 2>&1
+	/usr/bin/time -v clust2snp -i ${WD}/${READS1}.${READS2}.frc.fasta -n $N -m $M -x ${LCP} -y ${GSAtext} -z ${GSAsuff} > ${TIME_CLUST2SNP} 2>&1
 fi
 
 # 5.  Run clust2snp with parameter m -> reads1.reads2.frc.<m>.snp
 
 if [ ! -f ${WD}/${READS1}.${READS2}.frc.${M}.snp ]; then
 	echo "running clust2snp ..."
-	/usr/bin/time -v clust2snp -i ${WD}/${READS1}.${READS2}.frc.fasta -n $N -m $M > ${TIME_CLUST2SNP} 2>&1
+	/usr/bin/time -v clust2snp -i ${WD}/${READS1}.${READS2}.frc.fasta -n $N -m $M -x ${LCP} -y ${GSAtext} -z ${GSAsuff} > ${TIME_CLUST2SNP} 2>&1
 	mv ${WD}/${READS1}.${READS2}.frc.snp ${WD}/${READS1}.${READS2}.frc.${M}.snp
 fi
 
@@ -125,16 +134,34 @@ fi
 
 # 8.  Create reference of reads1.fasta using BWA MEM + bcftools + vcfconsensus
 
-if [ ! -f ${WD}/${READS1}.reference.fasta ]; then
-	echo "Computing reference of "${READS1}". Storing it to file "${WD}/${READS1}.reference.fasta" ..."
-	bwa mem ${WD}/${REF} ${WD}/${READS1}.fastq > ${WD}/alignment.tmp.sam
-	samtools view -b -S ${WD}/alignment.tmp.sam > ${WD}/alignment.tmp.bam
-	samtools sort ${WD}/alignment.tmp.bam > ${WD}/alignment.tmp.sorted.bam
-	samtools index ${WD}/alignment.tmp.sorted.bam
-	bcftools mpileup -f ${WD}/${REF} ${WD}/alignment.tmp.sorted.bam | bcftools call -mv -o ${WD}/calls.tmp.vcf
-	bgzip ${WD}/calls.tmp.vcf
-	vcf2fasta.sh ${WD}/calls.tmp.vcf.gz ${WD}/${REF} > ${WD}/${READS1}.reference.fasta
-	rm *.tmp*
+if [ "$samtool_version" -eq "1" ]; then
+	if [ ! -f ${WD}/${READS1}.reference.fasta ]; then
+		echo "Computing reference of "${READS1}". Storing it to file "${WD}/${READS1}.reference.fasta" ..."
+		bwa mem ${WD}/${REF} ${WD}/${READS1}.fastq > ${WD}/alignment.tmp.sam
+		samtools view -b -S ${WD}/alignment.tmp.sam > ${WD}/alignment.tmp.bam
+		samtools sort ${WD}/alignment.tmp.bam > ${WD}/alignment.tmp.sorted.bam
+		samtools index ${WD}/alignment.tmp.sorted.bam
+		bcftools mpileup -f ${WD}/${REF} ${WD}/alignment.tmp.sorted.bam | bcftools call -mv -o ${WD}/calls.tmp.vcf
+		bgzip ${WD}/calls.tmp.vcf
+		vcf2fasta.sh ${WD}/calls.tmp.vcf.gz ${WD}/${REF} > ${WD}/${READS1}.reference.fasta
+		rm *.tmp*
+	fi
+else
+	if [ ! -f ${WD}/${READS1}.reference.fasta ]; then
+		echo "Computing reference of "${READS1}". Storing it to file "${WD}/${READS1}.reference.fasta" ..."
+		bwa mem ${WD}/${REF} ${WD}/${READS1}.fastq > ${WD}/alignment.tmp.sam
+		samtools view -b -S ${WD}/alignment.tmp.sam > ${WD}/alignment.tmp.bam
+		samtools sort ${WD}/alignment.tmp.bam ${WD}/alignment.tmp.sorted
+		samtools index ${WD}/alignment.tmp.sorted.bam
+
+		#OLD SAMTOOLS/BCFTOOLS
+		samtools mpileup -uD -f ${WD}/${REF} ${WD}/alignment.tmp.sorted.bam > ${WD}/mpileup.tmp
+		bcftools view -vc ${WD}/mpileup.tmp > ${WD}/calls.tmp.vcf
+		
+		bgzip ${WD}/calls.tmp.vcf
+		vcf2fasta.sh ${WD}/calls.tmp.vcf.gz ${WD}/${REF} > ${WD}/${READS1}.reference.fasta
+		rm ${WD}/*.tmp*
+	fi
 fi
 
 # 9.  Builds BWA MEM index of reads1.reference.fasta -> reads1.reference.fasta.{amb,ann,bwt,fai,pac,sa} files
@@ -160,15 +187,32 @@ fi
 
 # 12. Generates VCF (bcftools calls) using BWA MEM + bcftools -> reads1.reads2.bcftools.vcf
 
-if [ ! -f ${WD}/${READS1}.${READS2}.bcftools.vcf ]; then
-	echo "Calling SNPs using BWA + bcftools. Storing SNPs to file "${WD}/${READS1}.${READS2}.bcftools.vcf" ..."
-	/usr/bin/time -v bwa mem ${WD}/${READS1}.reference.fasta ${WD}/${READS2}.fastq > ${WD}/alignment.tmp.sam 2> ${TIME_BWAMEM}
-	/usr/bin/time -v samtools view -b -S ${WD}/alignment.tmp.sam > ${WD}/alignment.tmp.bam 2>> ${TIME_BCFTOOLS}
-	/usr/bin/time -v samtools sort ${WD}/alignment.tmp.bam > ${WD}/alignment.tmp.sorted.bam 2>> ${TIME_BCFTOOLS}
-	/usr/bin/time -v samtools index ${WD}/alignment.tmp.sorted.bam 2>> ${TIME_BCFTOOLS}
-	/usr/bin/time -v bcftools mpileup -f ${WD}/${READS1}.reference.fasta ${WD}/alignment.tmp.sorted.bam | bcftools call -mv -o ${WD}/${READS1}.${READS2}.bcftools.vcf 2>> ${TIME_BCFTOOLS}
-	rm *.tmp*
+if [ "$samtool_version" -eq "1" ]; then
+	if [ ! -f ${WD}/${READS1}.${READS2}.bcftools.vcf ]; then
+		echo "Calling SNPs using BWA + bcftools. Storing SNPs to file "${WD}/${READS1}.${READS2}.bcftools.vcf" ..."
+		/usr/bin/time -v bwa mem ${WD}/${READS1}.reference.fasta ${WD}/${READS2}.fastq > ${WD}/alignment.tmp.sam 2> ${TIME_BWAMEM}
+		/usr/bin/time -v samtools view -b -S ${WD}/alignment.tmp.sam > ${WD}/alignment.tmp.bam 2>> ${TIME_BCFTOOLS}
+		/usr/bin/time -v samtools sort ${WD}/alignment.tmp.bam > ${WD}/alignment.tmp.sorted.bam 2>> ${TIME_BCFTOOLS}
+		/usr/bin/time -v samtools index ${WD}/alignment.tmp.sorted.bam 2>> ${TIME_BCFTOOLS}
+		/usr/bin/time -v bcftools mpileup -f ${WD}/${READS1}.reference.fasta ${WD}/alignment.tmp.sorted.bam | bcftools call -mv -o ${WD}/${READS1}.${READS2}.bcftools.vcf 2>> ${TIME_BCFTOOLS}
+		rm *.tmp*
+	fi
+else
+	if [ ! -f ${WD}/${READS1}.${READS2}.bcftools.vcf ]; then
+		echo "Calling SNPs using BWA + bcftools. Storing SNPs to file "${WD}/${READS1}.${READS2}.bcftools.vcf" ..."
+		/usr/bin/time -v bwa mem ${WD}/${READS1}.reference.fasta ${WD}/${READS2}.fastq > ${WD}/alignment.tmp.sam 2> ${TIME_BWAMEM}
+		/usr/bin/time -v samtools view -b -S ${WD}/alignment.tmp.sam > ${WD}/alignment.tmp.bam 2>> ${TIME_BCFTOOLS}
+		/usr/bin/time -v samtools sort ${WD}/alignment.tmp.bam ${WD}/alignment.tmp.sorted 2>> ${TIME_BCFTOOLS}
+		/usr/bin/time -v samtools index ${WD}/alignment.tmp.sorted.bam 2>> ${TIME_BCFTOOLS}
+
+		#OLD SAMTOOLS/BCFTOOLS
+		/usr/bin/time -v samtools mpileup -uD -f ${WD}/${READS1}.reference.fasta ${WD}/alignment.tmp.sorted.bam > ${WD}/mpileup.tmp 2>> ${TIME_BCFTOOLS}
+		/usr/bin/time -v bcftools view -vc ${WD}/mpileup.tmp > ${WD}/${READS1}.${READS2}.bcftools.vcf
+
+		rm ${WD}/*.tmp*
+	fi
 fi
+
 
 # 13. Generate report containing running times of eBWTclust pipeline, BWA+bcftools pipeline, and precision/recall of eBWTclust pipeline (using BWA+bcftools pipeline as ground truth)
